Attempting at code dedup
